# HarianWibuUI/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenApi specification - Alibou",
    "description": "OpenApi documentation for Spring Security",
    "termsOfService": "Terms of service",
    "contact": {
      "name": "Aji",
      "url": " https://some-url.com",
      "email": "contact@Aji.com"
    },
    "license": { "name": "Licence name", "url": "https://some-url.com" },
    "version": "1.0"
  },
  "servers": [
    { "url": "http://localhost:8088", "description": "Local ENV" },
    { "url":"https://some-url.com", "description": "PROD ENV" }
  ],
  "security": [{ "bearerAuth": [] }],
  "paths": {
    "/api/auth/signup": {
      "post": {
        "tags": ["auth-controller"],
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": { "*/*": { "schema": { "type": "object" } } }
          }
        }
      }
    },
    "/api/auth/authenticate": {
      "post": {
        "tags": ["auth-controller"],
        "operationId": "authenticate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AuthRequest" }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": { "$ref": "#/components/schemas/AuthResponse" }
              }
            }
          }
        }
      }
    },
    "/api/posts/{id}": {
      "get": {
        "tags": ["post-contoller"],
        "operationId": "viewAllPost",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "string" } } }
          }
        }
      }
    },
    "/api/posts/all": {
      "get": {
        "tags": ["post-contoller"],
        "operationId": "viewAllPost_1",
        "responses": {
          "200": {
            "description": "OK",
            "content": { "*/*": { "schema": { "type": "string" } } }
          }
        }
      }
    },
    "/api/auth/activate-account": {
      "get": {
        "tags": ["auth-controller"],
        "operationId": "confirm",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "OK" } }
      }
    }
  },
  "components": {
    "schemas": {
      "RegisterRequest": {
        "required": ["email", "password", "userName"],
        "type": "object",
        "properties": {
          "userName": { "type": "string" },
          "email": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "AuthRequest": {
        "required": ["email", "password"],
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "password": { "type": "string" }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": { "token": { "type": "string" } }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "JWT auth description",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}